detect area for each meetup, make north america fit the meetups
